Restructure data science repository
