+++ apps/spacex_dash/app.py
@@ -1,15 +1,19 @@
 # Import required libraries
-import pandas as pd
-import dash
-from dash import html
-from dash import dcc
-from dash.dependencies import Input, Output
-import plotly.express as px
+from pathlib import Path
 
-# Read the airline data into pandas dataframe
-spacex_df = pd.read_csv("spacex_launch_dash.csv")
-max_payload = spacex_df['Payload Mass (kg)'].max()
-min_payload = spacex_df['Payload Mass (kg)'].min()
+import dash
+import pandas as pd
+import plotly.express as px
+from dash import dcc, html
+from dash.dependencies import Input, Output
+
+
+DATA_PATH = Path(__file__).resolve().parents[2] / "data" / "processed" / "spacex_launch_dash.csv"
+
+# Read the SpaceX launch data into a pandas dataframe.
+spacex_df = pd.read_csv(DATA_PATH)
+max_payload = spacex_df["Payload Mass (kg)"].max()
+min_payload = spacex_df["Payload Mass (kg)"].min()
 
 # Create a dash application
 app = dash.Dash(__name__)
@@ -45,50 +49,49 @@ app.layout = html.Div(children=[html.H1('SpaceX Launch Records Dashboard',
                                 #dcc.RangeSlider(id='payload-slider',...)
                                 dcc.RangeSlider(id='payload-slider',
                                                 min=0, max=10000, step=1000,
-                                                marks={0: '0', 2500: '2500', 5000: '5000', 7500: '7500'}, 
+                                                marks={0: '0', 2500: '2500', 5000: '5000', 7500: '7500'},
                                                 value=[min_payload, max_payload]),
 
                                 # TASK 4: Add a scatter chart to show the correlation between payload and launch success
                                 html.Div(dcc.Graph(id='success-payload-scatter-chart')),
                                 ])
 
+
 # TASK 2:
 # Add a callback function for `site-dropdown` as input, `success-pie-chart` as output
 @app.callback(Output(component_id='success-pie-chart', component_property='figure'),
-                Input(component_id='site-dropdown', component_property='value'))
+              Input(component_id='site-dropdown', component_property='value'))
 def get_pie_chart(entered_site):
     if entered_site == 'ALL':
         fig = px.pie(spacex_df, values='class', names='Launch Site',
-                        title='Total Success Launches By Site')
+                     title='Total Success Launches By Site')
         return fig
-    else:
-        filtered_df = spacex_df[spacex_df['Launch Site'] == entered_site]
-        fig = px.pie(filtered_df, names='class',
-                        title=f'Total Success Launches for site {entered_site}')
-        return fig 
+    filtered_df = spacex_df[spacex_df['Launch Site'] == entered_site]
+    fig = px.pie(filtered_df, names='class',
+                 title=f'Total Success Launches for site {entered_site}')
+    return fig
 
 
 # TASK 4:
 # Add a callback function for `site-dropdown` and `payload-slider` as inputs, `success-payload-scatter-chart` as output
 @app.callback(Output(component_id='success-payload-scatter-chart', component_property='figure'),
-                [Input(component_id='site-dropdown', component_property='value'),
-                 Input(component_id='payload-slider', component_property='value')])
+              [Input(component_id='site-dropdown', component_property='value'),
+               Input(component_id='payload-slider', component_property='value')])
 def get_scatter_chart(entered_site, payload_range):
     low, high = payload_range
     mask = (spacex_df['Payload Mass (kg)'] >= low) & (spacex_df['Payload Mass (kg)'] <= high)
     filtered_df = spacex_df[mask]
-    
+
     if entered_site == 'ALL':
         fig = px.scatter(filtered_df, x='Payload Mass (kg)', y='class',
                          color='Booster Version Category',
                          title='Correlation between Payload and Success for All Sites')
         return fig
-    else:
-        filtered_df = filtered_df[filtered_df['Launch Site'] == entered_site]
-        fig = px.scatter(filtered_df, x='Payload Mass (kg)', y='class',
-                         color='Booster Version Category',
-                         title=f'Correlation between Payload and Success for site {entered_site}')
-        return fig
+    filtered_df = filtered_df[filtered_df['Launch Site'] == entered_site]
+    fig = px.scatter(filtered_df, x='Payload Mass (kg)', y='class',
+                     color='Booster Version Category',
+                     title=f'Correlation between Payload and Success for site {entered_site}')
+    return fig
 
 
 # Run the app
