+++ data/processed/spacex_launch_dash.csv
@@ -1,0 +1,57 @@
+,Flight Number,Launch Site,class,Payload Mass (kg),Booster Version,Booster Version Category
+0,1,CCAFS LC-40,0,0,F9 v1.0  B0003,v1.0
+1,2,CCAFS LC-40,0,0,F9 v1.0  B0004,v1.0
+2,3,CCAFS LC-40,0,525,F9 v1.0  B0005,v1.0
+3,4,CCAFS LC-40,0,500,F9 v1.0  B0006,v1.0
+4,5,CCAFS LC-40,0,677,F9 v1.0  B0007,v1.0
+5,7,CCAFS LC-40,0,3170,F9 v1.1,v1.1
+6,8,CCAFS LC-40,0,3325,F9 v1.1,v1.1
+7,9,CCAFS LC-40,0,2296,F9 v1.1,v1.1
+8,10,CCAFS LC-40,0,1316,F9 v1.1,v1.1
+9,11,CCAFS LC-40,0,4535,F9 v1.1,v1.1
+10,12,CCAFS LC-40,0,4428,F9 v1.1 B1011,v1.1
+11,13,CCAFS LC-40,0,2216,F9 v1.1 B1010,v1.1
+12,14,CCAFS LC-40,0,2395,F9 v1.1 B1012,v1.1
+13,15,CCAFS LC-40,0,570,F9 v1.1 B1013,v1.1
+14,16,CCAFS LC-40,0,4159,F9 v1.1 B1014,v1.1
+15,17,CCAFS LC-40,0,1898,F9 v1.1 B1015,v1.1
+16,18,CCAFS LC-40,0,4707,F9 v1.1 B1016,v1.1
+17,19,CCAFS LC-40,1,1952,F9 v1.1 B1018,v1.1
+18,20,CCAFS LC-40,1,2034,F9 FT B1019,FT
+19,22,CCAFS LC-40,0,5271,F9 FT B1020,FT
+20,23,CCAFS LC-40,1,3136,F9 FT B1021.1,FT
+21,24,CCAFS LC-40,1,4696,F9 FT B1022,FT
+22,25,CCAFS LC-40,1,3100,F9 FT B1023.1,FT
+23,26,CCAFS LC-40,0,3600,F9 FT B1024,FT
+24,27,CCAFS LC-40,1,2257,F9 FT B1025.1,FT
+25,28,CCAFS LC-40,1,4600,F9 FT B1026,FT
+26,6,VAFB SLC-4E,0,500,F9 v1.1  B1003,v1.1
+27,21,VAFB SLC-4E,0,553,F9 v1.1 B1017,v1.1
+28,29,VAFB SLC-4E,1,9600,F9 FT B1029.1,FT
+29,37,VAFB SLC-4E,1,9600,F9 FT B1036.1,FT
+30,40,VAFB SLC-4E,1,475,F9 FT B1038.1,FT
+31,42,VAFB SLC-4E,1,9600,F9 B4 B1041.1,B4
+32,46,VAFB SLC-4E,0,9600,F9 FT  B1036.2,FT
+33,49,VAFB SLC-4E,0,2150,F9 FT  B1038.2,FT
+34,51,VAFB SLC-4E,0,9600,F9 B4  B1041.2,B4
+35,55,VAFB SLC-4E,0,6460,F9 B4  B1043.2,B4
+36,30,KSC LC-39A,1,2490,F9 FT B1031.1,FT
+37,31,KSC LC-39A,0,5600,F9 FT B1030,FT
+38,32,KSC LC-39A,1,5300,F9 FT  B1021.2,FT
+39,33,KSC LC-39A,1,3696.65,F9 FT B1032.1,FT
+40,34,KSC LC-39A,0,6070,F9 FT B1034,FT
+41,35,KSC LC-39A,1,2708,F9 FT B1035.1,FT
+42,36,KSC LC-39A,1,3669,F9 FT  B1029.2,FT
+43,38,KSC LC-39A,0,6761,F9 FT B1037,FT
+44,39,KSC LC-39A,1,3310,F9 B4 B1039.1,B4
+45,41,KSC LC-39A,1,4990,F9 B4 B1040.1,B4
+46,43,KSC LC-39A,1,5200,F9 FT  B1031.2,FT
+47,44,KSC LC-39A,1,3500,F9 B4 B1042.1,B4
+48,54,KSC LC-39A,1,3600,F9 B5  B1046.1,B5
+49,45,CCAFS SLC-40,1,2205,F9 FT  B1035.2,FT
+50,47,CCAFS SLC-40,1,3696.65,F9 B4 B1043.1,B4
+51,48,CCAFS SLC-40,0,4230,F9 FT  B1032.2,FT
+52,50,CCAFS SLC-40,0,6092,F9 B4 B1044,B4
+53,52,CCAFS SLC-40,0,2647,F9 B4  B1039.2,B4
+54,53,CCAFS SLC-40,1,362,F9 B4 B1045.1,B4
+55,56,CCAFS SLC-40,0,5384,F9 B4  B1040.2,B4
