+++ data/processed/spacex_launch_geo.csv
@@ -1,0 +1,57 @@
+Flight Number,Date,Time (UTC),Booster Version,Launch Site,Payload,Payload Mass (kg),Orbit,Customer,Landing Outcome,class,Lat,Long
+1,2010-06-04,18:45:00,F9 v1.0  B0003,CCAFS LC-40,Dragon Spacecraft Qualification Unit,0,LEO,SpaceX,Failure   (parachute),0,28.56230197,-80.57735648
+2,2010-12-08,15:43:00,F9 v1.0  B0004,CCAFS LC-40,"Dragon demo flight C1, two CubeSats,  barrel of Brouère cheese",0,LEO (ISS),NASA (COTS) NRO,Failure   (parachute),0,28.56230197,-80.57735648
+3,2012-05-22,7:44:00,F9 v1.0  B0005,CCAFS LC-40,Dragon demo flight C2+,525,LEO (ISS),NASA (COTS),No attempt,0,28.56230197,-80.57735648
+4,2012-10-08,0:35:00,F9 v1.0  B0006,CCAFS LC-40,SpaceX CRS-1,500,LEO (ISS),NASA (CRS),No attempt,0,28.56230197,-80.57735648
+5,2013-03-01,15:10:00,F9 v1.0  B0007,CCAFS LC-40,SpaceX CRS-2,677,LEO (ISS),NASA (CRS),No attempt,0,28.56230197,-80.57735648
+7,2013-12-03,22:41:00,F9 v1.1,CCAFS LC-40,SES-8,3170,GTO,SES,No attempt,0,28.56230197,-80.57735648
+8,2014-01-06,22:06:00,F9 v1.1,CCAFS LC-40,Thaicom 6,3325,GTO,Thaicom,No attempt,0,28.56230197,-80.57735648
+9,2014-04-18,19:25:00,F9 v1.1,CCAFS LC-40,SpaceX CRS-3,2296,LEO (ISS),NASA (CRS),Controlled (ocean),0,28.56230197,-80.57735648
+10,2014-07-14,15:15:00,F9 v1.1,CCAFS LC-40,OG2 Mission 1  6 Orbcomm-OG2 satellites,1316,LEO,Orbcomm,Controlled (ocean),0,28.56230197,-80.57735648
+11,2014-08-05,8:00:00,F9 v1.1,CCAFS LC-40,AsiaSat 8,4535,GTO,AsiaSat,No attempt,0,28.56230197,-80.57735648
+12,2014-09-07,5:00:00,F9 v1.1 B1011,CCAFS LC-40,AsiaSat 6,4428,GTO,AsiaSat,No attempt,0,28.56230197,-80.57735648
+13,2014-09-21,5:52:00,F9 v1.1 B1010,CCAFS LC-40,SpaceX CRS-4,2216,LEO (ISS),NASA (CRS),Uncontrolled (ocean),0,28.56230197,-80.57735648
+14,2015-01-10,9:47:00,F9 v1.1 B1012,CCAFS LC-40,SpaceX CRS-5,2395,LEO (ISS),NASA (CRS),Failure  (drone ship),0,28.56230197,-80.57735648
+15,2015-02-11,23:03:00,F9 v1.1 B1013,CCAFS LC-40,DSCOVR,570,Sun–Earth L1,U.S. Air Force NASA NOAA,Controlled (ocean),0,28.56230197,-80.57735648
+16,2015-03-02,3:50:00,F9 v1.1 B1014,CCAFS LC-40,ABS-3A Eutelsat 115 West B,4159,GTO,ABS Eutelsat,No attempt,0,28.56230197,-80.57735648
+17,2015-04-14,20:10:00,F9 v1.1 B1015,CCAFS LC-40,SpaceX CRS-6,1898,LEO (ISS),NASA (CRS),Failure  (drone ship),0,28.56230197,-80.57735648
+18,2015-04-27,23:03:00,F9 v1.1 B1016,CCAFS LC-40,TürkmenÄlem 52°E / MonacoSAT,4707,GTO,Turkmenistan National Space Agency,No attempt,0,28.56230197,-80.57735648
+19,2015-06-28,14:21:00,F9 v1.1 B1018,CCAFS LC-40,SpaceX CRS-7,1952,LEO (ISS),NASA (CRS),Precluded  (drone ship),1,28.56230197,-80.57735648
+20,2015-12-22,1:29:00,F9 FT B1019,CCAFS LC-40,OG2 Mission 2   11 Orbcomm-OG2 satellites,2034,LEO,Orbcomm,Success  (ground pad),1,28.56230197,-80.57735648
+22,2016-03-04,23:35:00,F9 FT B1020,CCAFS LC-40,SES-9,5271,GTO,SES,Failure (drone ship),0,28.56230197,-80.57735648
+23,2016-04-08,20:43:00,F9 FT B1021.1,CCAFS LC-40,SpaceX CRS-8,3136,LEO (ISS),NASA (CRS),Success  (drone ship),1,28.56230197,-80.57735648
+24,2016-05-06,5:21:00,F9 FT B1022,CCAFS LC-40,JCSAT-14,4696,GTO,SKY Perfect JSAT Group,Success (drone ship),1,28.56230197,-80.57735648
+25,2016-05-27,21:39:00,F9 FT B1023.1,CCAFS LC-40,Thaicom 8,3100,GTO,Thaicom,Success  (drone ship),1,28.56230197,-80.57735648
+26,2016-06-15,14:29:00,F9 FT B1024,CCAFS LC-40,ABS-2A Eutelsat 117 West B,3600,GTO,ABS Eutelsat,Failure  (drone ship),0,28.56230197,-80.57735648
+27,2016-07-18,4:45:00,F9 FT B1025.1,CCAFS LC-40,SpaceX CRS-9,2257,LEO (ISS),NASA (CRS),Success (ground pad),1,28.56230197,-80.57735648
+28,2016-08-14,5:26:00,F9 FT B1026,CCAFS LC-40,JCSAT-16,4600,GTO,SKY Perfect JSAT Group,Success (drone ship),1,28.56230197,-80.57735648
+6,2013-09-29,16:00:00,F9 v1.1  B1003,VAFB SLC-4E,CASSIOPE,500,Polar LEO,MDA,Uncontrolled (ocean),0,34.63283416,-120.6107455
+21,2016-01-17,18:42:00,F9 v1.1 B1017,VAFB SLC-4E,Jason-3,553,LEO,NASA (LSP) NOAA CNES,Failure (drone ship),0,34.63283416,-120.6107455
+29,2017-01-14,17:54:00,F9 FT B1029.1,VAFB SLC-4E,Iridium NEXT 1,9600,Polar LEO,Iridium Communications,Success  (drone ship),1,34.63283416,-120.6107455
+37,2017-06-25,20:25:00,F9 FT B1036.1,VAFB SLC-4E,Iridium NEXT 2,9600,LEO,Iridium Communications,Success (drone ship),1,34.63283416,-120.6107455
+40,2017-08-24,18:51:00,F9 FT B1038.1,VAFB SLC-4E,Formosat-5,475,SSO,NSPO,Success (drone ship),1,34.63283416,-120.6107455
+42,2017-10-09,12:37:00,F9 B4 B1041.1,VAFB SLC-4E,Iridium NEXT 3,9600,Polar LEO,Iridium Communications,Success (drone ship),1,34.63283416,-120.6107455
+46,2017-12-23,1:27:00,F9 FT  B1036.2,VAFB SLC-4E,Iridium NEXT 4,9600,Polar LEO,Iridium Communications,Controlled (ocean),0,34.63283416,-120.6107455
+49,2018-02-22,14:17:00,F9 FT  B1038.2,VAFB SLC-4E,Paz  Tintin A & B,2150,SSO,Hisdesat exactEarth SpaceX,No attempt,0,34.63283416,-120.6107455
+51,2018-03-30,14:14:00,F9 B4  B1041.2,VAFB SLC-4E,Iridium NEXT 5,9600,Polar LEO,Iridium Communications,No attempt,0,34.63283416,-120.6107455
+55,2018-05-22,19:47:58,F9 B4  B1043.2,VAFB SLC-4E,"Iridium NEXT 6   GRACE-FO 1, 2",6460,Polar LEO,Iridium Communications GFZ • NASA,No attempt,0,34.63283416,-120.6107455
+30,2017-02-19,14:39:00,F9 FT B1031.1,KSC LC-39A,SpaceX CRS-10,2490,LEO (ISS),NASA (CRS),Success (ground pad),1,28.57325457,-80.64689529
+31,2017-03-16,6:00:00,F9 FT B1030,KSC LC-39A,EchoStar 23,5600,GTO,EchoStar,No attempt,0,28.57325457,-80.64689529
+32,2017-03-30,22:27:00,F9 FT  B1021.2,KSC LC-39A,SES-10,5300,GTO,SES,Success (drone ship),1,28.57325457,-80.64689529
+33,2017-05-01,11:15:00,F9 FT B1032.1,KSC LC-39A,NROL-76,3696.65,LEO,NRO,Success (ground pad),1,28.57325457,-80.64689529
+34,2017-05-15,23:21:00,F9 FT B1034,KSC LC-39A,Inmarsat-5 F4,6070,GTO,Inmarsat,No attempt,0,28.57325457,-80.64689529
+35,2017-06-03,21:07:00,F9 FT B1035.1,KSC LC-39A,SpaceX CRS-11,2708,LEO (ISS),NASA (CRS),Success (ground pad),1,28.57325457,-80.64689529
+36,2017-06-23,19:10:00,F9 FT  B1029.2,KSC LC-39A,BulgariaSat-1,3669,GTO,Bulsatcom,Success (drone ship),1,28.57325457,-80.64689529
+38,2017-07-05,23:38:00,F9 FT B1037,KSC LC-39A,Intelsat 35e,6761,GTO,Intelsat,No attempt,0,28.57325457,-80.64689529
+39,2017-08-14,16:31:00,F9 B4 B1039.1,KSC LC-39A,SpaceX CRS-12,3310,LEO (ISS),NASA (CRS),Success (ground pad),1,28.57325457,-80.64689529
+41,2017-09-07,14:00:00,F9 B4 B1040.1,KSC LC-39A,Boeing X-37B OTV-5,4990,LEO,U.S. Air Force,Success (ground pad),1,28.57325457,-80.64689529
+43,2017-10-11,22:53:00,F9 FT  B1031.2,KSC LC-39A,SES-11 / EchoStar 105,5200,GTO,SES EchoStar,Success (drone ship),1,28.57325457,-80.64689529
+44,2017-10-30,19:34:00,F9 B4 B1042.1,KSC LC-39A,Koreasat 5A,3500,GTO,KT Corporation,Success (drone ship),1,28.57325457,-80.64689529
+54,2018-05-11,20:14:00,F9 B5  B1046.1,KSC LC-39A,Bangabandhu-1,3600,GTO,Thales-Alenia/BTRC,Success (drone ship),1,28.57325457,-80.64689529
+45,2017-12-15,15:36:00,F9 FT  B1035.2,CCAFS SLC-40,SpaceX CRS-13,2205,LEO (ISS),NASA (CRS),Success (ground pad),1,28.56319718,-80.57682003
+47,2018-01-08,1:00:00,F9 B4 B1043.1,CCAFS SLC-40,Zuma,3696.65,LEO,Northrop Grumman,Success (ground pad),1,28.56319718,-80.57682003
+48,2018-01-31,21:25:00,F9 FT  B1032.2,CCAFS SLC-40,GovSat-1 / SES-16,4230,GTO,SES,Controlled (ocean),0,28.56319718,-80.57682003
+50,2018-03-06,5:33:00,F9 B4 B1044,CCAFS SLC-40,Hispasat 30W-6  PODSat,6092,GTO,Hispasat  NovaWurks,No attempt,0,28.56319718,-80.57682003
+52,2018-04-02,20:30:00,F9 B4  B1039.2,CCAFS SLC-40,SpaceX CRS-14,2647,LEO (ISS),NASA (CRS),No attempt,0,28.56319718,-80.57682003
+53,2018-04-18,22:51:00,F9 B4 B1045.1,CCAFS SLC-40,Transiting Exoplanet Survey Satellite (TESS),362,HEO,NASA (LSP),Success (drone ship),1,28.56319718,-80.57682003
+56,2018-06-04,4:45:00,F9 B4  B1040.2,CCAFS SLC-40,SES-12,5384,GTO,SES,No attempt,0,28.56319718,-80.57682003
