+++ scripts/childtest.py
