+++ src/testrepo/__init__.py
@@ -1,0 +1,1 @@
+"""Reusable helpers for the testrepo data science projects."""
